first steps toward centralized credential management
The end goal is ensuring that no code outside of UserCredentials
manipulates `users#current_login`.

A reoccuring pattern I've found is changing the value of
`users#current_login` to access the data of another user.
This seems somewhat suboptimal and perhaps error-prone,
so I hope it can be refactored away by adjusting the Storage
internals later on.